test(daemon): split owner wait regressions
